requirements.txt

fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.9.2
cryptography==43.0.1
pycryptodome>=3.20.0